Add integration tests for authentication
