+++ test/stream.test.js
@@ -6,7 +6,6 @@ import Stream from '../src/js/stream';
 import renderComments from './methods/stream/render-comments';
 import init from './methods/stream/init';
 import login from './methods/stream/login';
-import getJsonWebToken from './methods/stream/get-json-web-token';
 import publishEvent from './methods/stream/publish-event';
 
 describe("Stream", () => {
@@ -17,6 +16,5 @@ describe("Stream", () => {
 	describe('.init', init);
 	describe('.renderComments', renderComments);
 	describe('.login', login);
-	describe('.getJsonWebToken', getJsonWebToken);
 	describe('.publishEvent', publishEvent);
 });
